feat(product): six estate preferences (fact_extraction, consolidation, contradiction_sweep, cross_encoder_routing, maintenance, adaptive_recall) seeded by format 1.9 and set from the mootx01 preference command; consolidation, contradiction sweep, temporal causality, training and the end-of-day tournament run live behind them; the maintenance engine keeps three checks and runs on its signals; similar_recall over the LSA lane ships as moot_recall_similar (81 operations, digest repinned), both ports
Ported-from: eb26de473275533c821a83fa91669de3c98c5f05

+++ apps/mootx01/Tests/MootInstallerCoreTests/UpgradeReclaimSetTests.swift
@@ -49,7 +49,7 @@ struct UpgradeReclaimSetTests {
         for value in values {
             withUnsafeBytes(of: value.bitPattern.littleEndian) { bytes.append(contentsOf: $0) }
         }
-        return VectorPayload(kind: .float32, dim: values.count, bytes: bytes)
+        return VectorPayload(kind: .float32, dim: UInt32(values.count), bytes: bytes)
     }
 
     /// Every `model_id` left in the `vectors` table, one entry per row.
